requirements.txt
flask==3.0.2
flask-cors==4.0.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
ultralytics==8.1.27
torch==2.2.0
pillow==10.2.0
numpy==1.26.4 